Fix: 게시글 상세 조회 시 옵션 정보 조회 오류 수정

+++ src/main/java/world/startoy/polling/usecase/dto/PollOptionResponse.java
@@ -18,7 +18,7 @@ public class PollOptionResponse {
     @NotNull
     private String pollOptionText;
     @NotNull
-    private int votedCount; // 득표수
+    private Long votedCount; // 득표수 // JPQL 쿼리에서 COUNT(v.id)는 Long 타입을 반환
     @NotNull
     private String fileUid;
     @NotNull
